reorganize more pharmacy data into PharmacySelectionProvider
