+++ packages/components/src/systems/PrescribeProvider/PrescribeProvider.tsx
@@ -2,35 +2,21 @@ import {
   Accessor,
   createContext,
   createEffect,
-  createMemo,
   createSignal,
   JSXElement,
   useContext
 } from 'solid-js';
 import { usePhotonClient } from '../SDKProvider';
 import { Prescription } from '@photonhealth/sdk/dist/types';
-import {
-  GenerateCoverageOptions,
-  GetPatientPreferredPharmaciesAndAddress,
-  GetPharmaciesQuery
-} from '../../fetch';
+import { GenerateCoverageOptions, GetPatientPreferredPharmaciesAndAddress } from '../../fetch';
 import { triggerToast } from '../../index';
 import { useDraftPrescriptions } from '../DraftPrescriptions';
-import {
-  combineAllRoutingConstraints,
-  getRoutingConstraint,
-  RoutingConstraint
-} from '../RoutingConstraints';
-import getLocation from '../../utils/getLocations';
-import { useGoogleService } from '../GoogleServiceProvider';
+import { usePharmacySelectionContext, Address } from '../PharmacySelect/PharmacySelectionProvider';
 import { Env } from '@photonhealth/sdk';
 
 export type PrescribeContextType = {
   // values
   coverageOptions: Accessor<CoverageOption[]>;
-  routingConstraints: Accessor<RoutingConstraint[]>;
-  combinedRoutingConstraint: Accessor<RoutingConstraint>;
-  unroutablePharmacyIds: Accessor<Set<string>>;
   selectedCoverageOption: Accessor<CoverageOption | undefined>;
 
   // actions
@@ -66,17 +52,9 @@ export type CoverageOption = {
   pharmacy: { id: string; name: string };
 };
 
-export type Address = {
-  street1: string;
-  street2: string;
-  city: string;
-  state: string;
-  postalCode: string;
-};
-
 export const PrescribeProvider = (props: PrescribeProviderProps) => {
-  const { googleMapsServices } = useGoogleService();
   const { draftPrescriptions, setDraftPrescriptions } = useDraftPrescriptions();
+  const pharmacySelectionContext = usePharmacySelectionContext();
   const client = usePhotonClient();
   const [coverageOptions, setCoverageOptions] = createSignal<CoverageOption[]>([]);
   const [patientPreferredPharmacyId, setPatientPreferredPharmacyId] = createSignal<string | null>(
@@ -89,51 +67,6 @@ export const PrescribeProvider = (props: PrescribeProviderProps) => {
   const [selectedCoverageOption, setSelectedCoverageOption] = createSignal<
     CoverageOption | undefined
   >();
-
-  const routingConstraints = createMemo((): RoutingConstraint[] => {
-    return draftPrescriptions().map((prescription: Prescription) =>
-      getRoutingConstraint(prescription)
-    );
-  });
-
-  const combinedRoutingConstraint = createMemo(() => {
-    return combineAllRoutingConstraints(routingConstraints());
-  });
-
-  const unroutablePharmacyIds = createMemo(() => {
-    const combinedExcludeRoutingConstraint = combineAllRoutingConstraints(routingConstraints(), [
-      'exclude'
-    ]);
-    return new Set(
-      combinedExcludeRoutingConstraint.constraint_pharmacies?.map((pharmacy) => pharmacy.id) || []
-    );
-  });
-
-  async function fetchPharmacies(location: { latitude: number; longitude: number }) {
-    const { data } = await client!.apollo.query({
-      query: GetPharmaciesQuery,
-      variables: {
-        location: { latitude: location?.latitude, longitude: location?.longitude }
-      }
-    });
-    if (!data?.pharmacies) {
-      return [];
-    }
-
-    return data.pharmacies;
-  }
-
-  async function fetchLocalPharmacies(address: Address) {
-    const { geocoder } = googleMapsServices();
-    if (!geocoder) throw new Error('Geocoder not loaded');
-
-    const stringAddress = `${address?.street1}, ${address?.city}, ${address?.state} ${address?.postalCode}`;
-
-    const locations = await getLocation(stringAddress, geocoder);
-
-    const pharmacies = await fetchPharmacies(locations[0]);
-    return pharmacies;
-  }
 
   createEffect(() => {
     if (props.patientId) {
@@ -190,7 +123,7 @@ export const PrescribeProvider = (props: PrescribeProviderProps) => {
         // check coverage on a walgreens or cvs pharmacy near the patient
         const address = patientAddress();
         if (address) {
-          fetchLocalPharmacies(address).then((pharmacies) => {
+          pharmacySelectionContext.fetchLocalPharmacies(address).then((pharmacies) => {
             let localPharmacyId: string | null = null;
             const majorChainPharmacy = pharmacies.find(
               (pharmacy: { id: string; name: string }) =>
@@ -254,17 +187,16 @@ export const PrescribeProvider = (props: PrescribeProviderProps) => {
     return response.data.generateCoverageOptions as CoverageOption[];
   };
 
+  // Coordinator: when a coverage option is selected, sync pharmacyId to the pharmacy provider
   const selectOtherCoverageOption = (value: CoverageOption) => {
     setSelectedCoverageOption(value);
     setDidSelectOtherCoverageOption(true);
+    pharmacySelectionContext.setPharmacyId(value.pharmacy.id);
   };
 
   const value = {
     // values
     coverageOptions,
-    routingConstraints,
-    combinedRoutingConstraint,
-    unroutablePharmacyIds,
     selectedCoverageOption,
 
     // actions
