+++ packages/components/src/systems/TestMocks/MockPrescribeProvider.tsx
@@ -9,15 +9,6 @@ export const mockPrescribeContextValues = () => {
   return {
     // mock values
     coverageOptions: () => [],
-    routingConstraints: () => [],
-    combinedRoutingConstraint: () => {
-      return {
-        prescriptions: [],
-        routing_constraint_type: 'exclude',
-        constraint_pharmacies: []
-      };
-    },
-    unroutablePharmacyIds: () => new Set(),
     selectedCoverageOption: () => undefined,
     // mock actions
     selectOtherCoverageOption: vi.fn()
